Adding Leave Balace while we create new customer

+++ Holidayview.Domain/Model/Vacation.cs
@@ -5,7 +5,7 @@ namespace Holidayview.Domain.Model
     public class Vacation
     {
         public int Id { get; set; }
-        public int LeaveDimension { get; set; }
+        public double LeaveDimension { get; set; }
 
         //Customer
         public virtual ICollection<Customer> Customers { get; set; }
